pyproject.toml
[project]
name = "escape-room"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "thelegendofmario", email = "ethanfrancis577@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
]

[project.scripts]
escape-room = "escape_room.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
